update unit test

+++ tests/test_models/test_losses/test_inverseform_loss.py
@@ -1,4 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
+
 import os
 
 import pytest
@@ -6,10 +7,10 @@ import requests
 import torch
 from tqdm import tqdm
 
-from mmseg.models.losses import InverseFormLoss
-
 
 def test_inverseform_loss():
+    from mmseg.models.losses import InverseFormLoss
+
     # construct test data
     pred = torch.zeros(3, 1, 4, 4)
     target = torch.zeros(3, 1, 4, 4)
@@ -26,25 +27,30 @@ def test_inverseform_loss():
     print(pretraind_model_url)
     os.makedirs('./checkpoints', exist_ok=True)
     inverseNet_path = './checkpoints/distance_measures_regressor.pth'
-
-    response = requests.get(pretraind_model_url, stream=True)
-    total_size_in_bytes = int(response.headers.get('content-length', 0))
-    block_size = 1024  # 1 Kibibyte
-    progress_bar = tqdm(total=total_size_in_bytes, unit='iB', unit_scale=True)
-    with open(inverseNet_path, 'wb') as file:
-        for data in response.iter_content(block_size):
-            progress_bar.update(len(data))
-            file.write(data)
-    progress_bar.close()
-    if total_size_in_bytes != 0 and progress_bar.n != total_size_in_bytes:
-        print('ERROR, something went wrong')
+    if not os.path.exists(inverseNet_path):
+        response = requests.get(pretraind_model_url, stream=True)
+        total_size_in_bytes = int(response.headers.get('content-length', 0))
+        block_size = 1024  # 1 Kibibyte
+        progress_bar = tqdm(
+            total=total_size_in_bytes, unit='iB', unit_scale=True)
+        with open(inverseNet_path, 'wb') as file:
+            for data in response.iter_content(block_size):
+                progress_bar.update(len(data))
+                file.write(data)
+        progress_bar.close()
+        if total_size_in_bytes != 0 and progress_bar.n != total_size_in_bytes:
+            print('ERROR, something went wrong')
 
     # test inverseform_loss
-    loss_class = InverseFormLoss(inverseNet_path=inverseNet_path)
-    with pytest.raises(AssertionError):
-        loss = loss_class(pred, target)
+    loss_class = InverseFormLoss(
+        inverseNet_path=inverseNet_path, map_location='cpu')
+    loss = loss_class(pred, target)
+    print(loss)
     assert isinstance(loss, torch.Tensor)
 
-
-# if __name__ == "__main__":
-#     test_inverseform_loss()
+    # Test loss forward with pred.shape!=target.shape
+    pred_wrong = torch.zeros(2, 1, 4, 4)
+    target_wrong = torch.zeros(3, 1, 4, 4)
+    with pytest.raises(AssertionError):
+        loss = loss_class(pred_wrong, target_wrong)
+        print(loss)
